third_party/blink/web_tests/platform/generic/external/wpt/custom-elements/perform-microtask-checkpoint-before-construction-xml-parser-expected.txt

This is a testharness.js-based test.
FAIL XML parser must perform a microtask checkpoint before constructing a custom element assert_equals: expected 1 but got 0
Harness: the test ran to completion.
